+++ docs/source/changelog.rst
@@ -8,6 +8,7 @@ Enhancements
 ------------
 
 - Added :meth:`dask_ml.preprocessing.RobustScaler` (:issue:`62`)
+- Added :meth:`dask_ml.preprocessing.OrdinalEncoder` (:issue:`119`)
 
 
 Version 0.4.0